backend/requirements.txt
asgiref
Django
django-cors-headers
django-csp
djangorestframework
djangorestframework-simplejwt
gunicorn
Pillow
psycopg2-binary
PyJWT
pytz
sqlparse
whitenoise
boto3
django-storages
selenium
